requirements.txt
numpy==2.1.2
python-dotenv==1.0.1
livekit==0.17.5
livekit-agents==0.10.2
pydub==0.25.1
livekit-plugins-silero==0.7.2
livekit-plugins-cartesia==0.4.2
livekit-plugins-deepgram==0.6.8
livekit-plugins-openai==0.10.4